+++ CHANGELOG.md
@@ -4,6 +4,8 @@ All notable changes to TTT2 will be documented here. Inspired by [keep a changel
 
 ## Unreleased
 
+## [v0.11.1b](https://github.com/TTT-2/TTT2/tree/v0.11.1b) (2021-11-16)
+
 ### Added
 
 - Added four new Karma multipliers as role variables. They are applied **after** all other Karma calculations are done_